+++ plugins/postgres/dbt/include/postgres/macros/adapters.sql
@@ -14,21 +14,21 @@
   );
 {%- endmacro %}
 
-{% macro postgres__create_schema(database_name, schema_name) -%}
-  {% if database_name -%}
-    {{ adapter.verify_database(database_name) }}
+{% macro postgres__create_schema(relation) -%}
+  {% if relation.database -%}
+    {{ adapter.verify_database(relation.database) }}
   {%- endif -%}
   {%- call statement('create_schema') -%}
-    create schema if not exists {{ schema_name }}
+    create schema if not exists {{ relation.without_identifier().include(database=False) }}
   {%- endcall -%}
 {% endmacro %}
 
-{% macro postgres__drop_schema(database_name, schema_name) -%}
-  {% if database_name -%}
-    {{ adapter.verify_database(database_name) }}
+{% macro postgres__drop_schema(relation) -%}
+  {% if relation.database -%}
+    {{ adapter.verify_database(relation.database) }}
   {%- endif -%}
   {%- call statement('drop_schema') -%}
-    drop schema if exists {{ schema_name }} cascade
+    drop schema if exists {{ relation.without_identifier().include(database=False) }} cascade
   {%- endcall -%}
 {% endmacro %}
 
@@ -54,23 +54,23 @@
 {% endmacro %}
 
 
-{% macro postgres__list_relations_without_caching(information_schema, schema) %}
+{% macro postgres__list_relations_without_caching(schema_relation) %}
   {% call statement('list_relations_without_caching', fetch_result=True) -%}
     select
-      '{{ information_schema.database }}' as database,
+      '{{ schema_relation.database }}' as database,
       tablename as name,
       schemaname as schema,
       'table' as type
     from pg_tables
-    where schemaname ilike '{{ schema }}'
+    where schemaname ilike '{{ schema_relation.schema }}'
     union all
     select
-      '{{ information_schema.database }}' as database,
+      '{{ schema_relation.database }}' as database,
       viewname as name,
       schemaname as schema,
       'view' as type
     from pg_views
-    where schemaname ilike '{{ schema }}'
+    where schemaname ilike '{{ schema_relation.schema }}'
   {% endcall %}
   {{ return(load_result('list_relations_without_caching').table) }}
 {% endmacro %}